поправил там верстку туда сюда

+++ JustSimpleShop/models.py
@@ -1,3 +1,4 @@
+
 from django.db import models
 
 # Create your models here.
@@ -27,7 +28,7 @@ class Product(models.Model):
     name = models.CharField(max_length=100, db_index=True)
     slug = models.SlugField(max_length=100, db_index=True)
     description = models.TextField(blank=True)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.PositiveIntegerField()
     available = models.BooleanField(default=True)
     stock = models.PositiveIntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
